refactor: overall framework
