+++ .gitignore
@@ -16,3 +16,5 @@ Cargo.lock
 
 assets/
 test/
+
+.vscode/
