+++ llm_apis/src/chat/oneonone.rs
@@ -1,0 +1,17 @@
+use kalosm::language::*;
+
+pub async fn communicate() {
+    let mut model = Llama::new_chat();
+    let mut chat = Chat::builder(&mut model)
+        .with_system_prompt("The assistant help answer questions based on the context given by the user. The model knows that the information the user gives it is always true.")
+        .build();
+
+    loop {
+        chat.add_message(prompt_input("\n> ").unwrap())
+            .await
+            .unwrap()
+            .to_std_out()
+            .await
+            .unwrap();
+    }
+}
